more test => less error

+++ ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: nalebrun <nalebrun@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/14 14:00:41 by nalebrun          #+#    #+#             */
-/*   Updated: 2024/10/22 09:28:11 by nalebrun         ###   ########.fr       */
+/*   Updated: 2024/10/22 16:24:58 by nalebrun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,9 +23,9 @@ static int	ft_isspace(char c)
 
 int	ft_atoi(const char *str)
 {
-	int	i;
-	int	sign;
-	int	res;
+	long long int	i;
+	long long int	sign;
+	long long int	res;
 
 	res = 0;
 	sign = 1;
